src/rag/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
colorama>=0.4.6
faiss-cpu>=1.7.4
numpy>=1.21.0
tqdm>=4.65.0
rich>=13.0.0
tiktoken>=0.5.0
halo>=0.0.31
elevenlabs>=0.2.24
SpeechRecognition>=3.8.1
pyaudio>=0.2.11
keyboard>=0.13.5
pyttsx3>=2.90
pocketsphinx>=5.0.0  # For offline STT fallback 